chore: 생성 페이지 수 압축
