+++ frontend/src/pages/Create/Steps/Step5/index.styled.ts
@@ -11,24 +11,35 @@ const Label = styled.div`
   display: flex;
   justify-content: flex-end;
 
-  max-width: 40rem;
+  width: 100%;
 `;
 
 const TextArea = styled.textarea`
-  width: 90%;
-  max-width: 40rem;
+  resize: vertical;
+
+  width: 100%;
   height: 15rem;
 
   border-radius: 0.5rem;
 
+  box-sizing: border-box;
   padding: 0.5rem;
 
   font-size: 1.2rem;
 
-  ${({ theme: { colors } }) => `
+  ${({
+    theme: {
+      colors,
+      breakpoints: { md },
+    },
+  }) => `
     background: ${colors.white001};
 
     border: 1px solid ${colors.gray003};
+
+    @media only screen and (max-width: ${md}px) {
+      font-size: 1rem;
+    }
   `}
 `;
