+++ frontend/src/pages/Create/Steps/index.ts
@@ -3,6 +3,3 @@ export { default as Step2 } from './Step2';
 export { default as Step3 } from './Step3';
 export { default as Step4 } from './Step4';
 export { default as Step5 } from './Step5';
-export { default as Step6 } from './Step6';
-export { default as Step7 } from './Step7';
-export { default as Step8 } from './Step8';
